+++ geotrek/tourism/models.py
@@ -203,7 +203,7 @@ class TouristicContentCategory(PictogramMixin):
         return self.label
 
 
-class TouristicContentType(models.Model):
+class TouristicContentType(PictogramMixin):
 
     label = models.CharField(verbose_name=_(u"Label"), max_length=128, db_column='nom')
     category = models.ForeignKey(TouristicContentCategory, related_name='types',
@@ -328,7 +328,7 @@ TouristicContent.add_property('touristic_contents', lambda self: intersecting(To
 TouristicContent.add_property('published_touristic_contents', lambda self: intersecting(TouristicContent, self, distance=settings.TOURISM_INTERSECTION_MARGIN).filter(published=True), _(u"Published touristic contents"))
 
 
-class TouristicEventType(models.Model):
+class TouristicEventType(PictogramMixin):
 
     type = models.CharField(verbose_name=_(u"Type"), max_length=128, db_column='type')